CFlow to swift Publisher
